gemini_chatbot/requirements.txt
google-genai>=0.4.0
python-dotenv>=1.0.0
rich>=13.3.0
typer>=0.9.0 